refactor: Refcatored AspectModelParquetPayloadGenerator class
